# Conflicts:
#	AuthUsersMicroService/src/main/java/com/example/authusersmicroservice/AuthUsersMicroServiceApplication.java

+++ InterviewMicroService/src/main/java/com/example/interviewmicroservice/models/Appointment.java
@@ -17,7 +17,7 @@ public class Appointment {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long appointmentId;
+    private Integer appointmentId;
 
     @Column(nullable = false)
     private String clientEmail;
